Reorder aliases a bit

+++ home/zsh/aliases.zsh
@@ -3,12 +3,6 @@ if [[ `uname -s` != "OpenBSD" ]]; then
 	alias ls='ls -G'
 fi
 
-if [[ `uname -s` == "Darwin" ]]; then
-    alias pubkeycp='cat ~/.ssh/id_rsa.pub |pbcopy |echo "Public key copied to clipboard."'
-    alias hibernateon='sudo pmset -a hibernatemode 5'
-    alias hibernateoff='sudo pmset -a hibernatemode 0'
-fi
-
 # A virtualenv is required in order to install python packages (see export
 # PIP_REQUIRE_VIRTUALENV above), but sometimes I want to install or upgrade
 # global packages. That's where I use gpip.
@@ -42,17 +36,23 @@ alias pushcapiohome="rsync -arP --delete-after /Users/sjk/Documents/Capio/ncsvak
 
 alias burniso='cdrecord -v -eject dev=1,0,0'
 
-# OS X convenience aliases
-alias lockmac='/System/Library/CoreServices/Menu\ Extras/User.menu/Contents/Resources/CGSession -suspend'
-alias stfu='osascript -e "set volume output muted true"'
-alias v1='osascript -e "set volume 1"'
-alias v3='osascript -e "set volume 3"'
-alias v5='osascript -e "set volume 5"'
-alias v7='osascript -e "set volume 7"'
-alias v9='osascript -e "set volume 9"'
-alias hidedesktop='defaults write com.apple.finder CreateDesktop -bool false && killall Finder'
-alias showdesktop='defaults write com.apple.finder CreateDesktop -bool true && killall Finder'
 alias thesession='timidity ~/Downloads/*.mid(Nm0om[1])'
+
+# OS X convenience aliases
+if [[ `uname -s` == "Darwin" ]]; then
+    alias pubkeycp='cat ~/.ssh/id_rsa.pub |pbcopy |echo "Public key copied to clipboard."'
+    alias hibernateon='sudo pmset -a hibernatemode 5'
+    alias hibernateoff='sudo pmset -a hibernatemode 0'
+    alias lockmac='/System/Library/CoreServices/Menu\ Extras/User.menu/Contents/Resources/CGSession -suspend'
+    alias stfu='osascript -e "set volume output muted true"'
+    alias v1='osascript -e "set volume 1"'
+    alias v3='osascript -e "set volume 3"'
+    alias v5='osascript -e "set volume 5"'
+    alias v7='osascript -e "set volume 7"'
+    alias v9='osascript -e "set volume 9"'
+    alias hidedesktop='defaults write com.apple.finder CreateDesktop -bool false && killall Finder'
+    alias showdesktop='defaults write com.apple.finder CreateDesktop -bool true && killall Finder'
+fi
 
 # Make the Terminal icon bounce. Sometimes useful, like:
 # execute_timey_wimey_command && termbounce
